fix after code review #4
