+++ src/app/myProfile/MyProfile.tsx
@@ -1,17 +1,13 @@
 import { useState } from 'react';
 import './MyProfile.scss';
 import ProfileRead from './profileRead/ProfileRead';
-import ProfileUpdate, { Profile } from './profileUpdate/ProfileUpdate';
+import ProfileUpdate from './profileUpdate/ProfileUpdate';
+import { Profile } from '@/types/entities/UserProfile';
+import { mockUserProfile } from './mockUserProfile';
 
 const MyProfile = () => {
   const [isEditing, setIsEditing] = useState(false);
-  const [profile, setProfile] = useState<Profile>({
-    firstname: 'John',
-    lastname: 'Doe',
-    username: 'john.doe.gym@gmail.com',
-    location: 'Los Angeles, CA',
-    bio: 'The most dangerous. A beast slayer. A conqueror of conqueror, the goat of all goats. The ultimate needle mover, the head of the table, the tribal chief. In god mode himself',
-  });
+  const [profile, setProfile] = useState<Profile>(mockUserProfile);
 
   const toggleIsEditing = () => {
     setIsEditing(!isEditing);
@@ -24,8 +20,11 @@ const MyProfile = () => {
 
   return (
     <div className='my_profile'>
-      {!isEditing && <ProfileRead editAction={toggleIsEditing} profile={profile} />}
-      {isEditing && <ProfileUpdate cancelAction={toggleIsEditing} profile={profile} saveAction={handleSaveProfile} />}
+      {isEditing ? (
+        <ProfileUpdate cancelAction={toggleIsEditing} profile={profile} saveAction={handleSaveProfile} />
+      ) : (
+        <ProfileRead editAction={toggleIsEditing} profile={profile} />
+      )}
     </div>
   );
 };
