+++ src/app/myProfile/mockUserProfile.ts
@@ -1,0 +1,8 @@
+export const mockUserProfile = {
+  firstname: 'John',
+  lastname: 'Doe',
+  username: 'john.doe.gym@gmail.com',
+  location: 'Los Angeles, CA',
+  bio: 'The most dangerous. A beast slayer. A conqueror of conqueror, the goat of all goats. The ultimate needle mover, the head of the table, the tribal chief. In god mode himself',
+  avatarUrl: 'https://www.fit.pl/img/2007/02/_max/arnie.jpg',
+};
