+++ src/types/entities/UserProfile.ts
@@ -1,0 +1,8 @@
+export interface Profile {
+  firstname: string;
+  lastname: string;
+  username: string;
+  bio: string;
+  avatarUrl: string;
+  location: string;
+}
